+++ HISTORY.rst
@@ -35,6 +35,12 @@ Changes from API 2019-02-19:
 	- Added ``Account.branding_logo`` as a ForeignKey
     - Populate ``Account.branding_icon`` and ``.branding_logo`` from the new ``Account.settings.branding.icon`` and ``.logo``
 
+Changes from API 2019-03-14:
+
+- Renamed ``Invoice.application_fee`` to ``Invoice.application_fee_amount`` (added deprecated property for the old name)
+- Removed ``Invoice.date``, in place of ``Invoice.created`` (added deprecated property for the old name)
+- Added ``Invoice.status_transitions``
+
 2.0.3 (2019-06-11)
 ------------------